backend/requirements.txt
Django
djangorestframework
django-cors-headers
channels
channels-redis
celery
redis
psycopg2-binary
django-timescaledb
pgvector
cryptography
numpy
pandas
scikit-learn
sentence-transformers
langchain
langchain-openai
httpx
python-dotenv
gunicorn
uvicorn